Rely on xplugd for detecting dock/undock events
This patch adds support for detecting when a monitor is plugged in or
unplugged.  Useful when docking or when starting a presentation.

Included xplugd.script is called by xplugd on RandR events to X.  See
https://github.com/troglobit/xplugd for details on xplugd.

Also, a new randr.lua module is added in this patch.  It takes care to
call the xplugd.script on initial login, i.e., when starting awesome on
a computer already connected to all screens.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>